frontend2: clean up num validation behavior

+++ packages/frontend2/src/components/form-element.ts
@@ -18,7 +18,7 @@ export interface FormElementProps {
   prefix?: string;
   suffix?: string;
   validation?: {
-    type: "numeric";
+    type: "int";
     min?: number;
     max?: number;
   };
